modules/crawler/requirements.txt
beautifulsoup4==4.9.0
certifi==2020.4.5.1
chardet==3.0.4
html5lib==1.0.1
idna==2.9
pymongo==3.10.1
PyYAML==5.3.1
requests==2.23.0
six==1.14.0
soupsieve==2.0
urllib3==1.25.9
webencodings==0.5.1
pylint===2.4.4
yapf===0.30.0